toolkit/content/tests/widgets/test_menulist_null_value.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="/tests/SimpleTest/test.css" type="text/css"?>

<window title="Menulist value property"
  onload="setTimeout(runTests, 0);"
  xmlns:html="http://www.w3.org/1999/xhtml"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <title>Menulist value property</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>      
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>      

<menulist id="list">
  <menupopup>
    <menuitem id="i1" label="One" value="item1"/>
    <menuitem id="i2" label="Two" value="item2"/>
  </menupopup>
</menulist>

<script class="testbody" type="application/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

function runTests()
{
  var list = document.getElementById("list");

  list.value = "item2";
  is(list.value, "item2", "Check list value after setting value");
  is(list.getAttribute("label"), "Two", "Check list label after setting value");

  list.selectedItem = null;
  is(list.value, "", "Check list value after setting selectedItem to null");
  is(list.getAttribute("label"), "", "Check list label after setting selectedItem to null");

  // select something again to make sure the label is not already empty
  list.selectedIndex = 0;
  is(list.value, "item1", "Check list value after setting selectedIndex");
  is(list.getAttribute("label"), "One", "Check list label after setting selectedIndex");

  // set the value to null and test it (bug 408940)
  list.value = null;
  is(list.value, "", "Check list value after setting value to null");
  is(list.getAttribute("label"), "", "Check list label after setting value to null");

  SimpleTest.finish();
}

]]>
</script>

<body xmlns="http://www.w3.org/1999/xhtml">
<p id="display">
</p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
</body>

</window>
